factor out queue creation since we need to reuse it
